first_agent/pyproject.toml
[project]
name = "helpful-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.24.0",
    "openai-agents>=0.10.2",
    "python-dotenv>=1.2.1",
]
